Move emojiMod helper out into helpers/modLog
